akhavan/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-ckeditor-5==0.2.15
django-jalali==7.2.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.0.0
phonenumbers==8.13.50
pillow==11.0.0
PyYAML==6.0.2
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
